# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swasthya Card API",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {}
  },
  "paths": {
    "/auth/otp/request": {
      "post": {
        "summary": "Request an OTP",
        "tags": [
          "auth"
        ],
        "description": "Starts login/registration. REQ-AUTH-002/003.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "pattern": "^\\+[1-9]\\d{6,14}$"
                  }
                },
                "required": [
                  "phone"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "otpSentTo": {
                          "type": "string"
                        },
                        "expiresInSec": {
                          "type": "number"
                        },
                        "demoOtp": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "otpSentTo",
                        "expiresInSec"
                      ],
                      "additionalProperties": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/otp/verify": {
      "post": {
        "summary": "Verify an OTP",
        "tags": [
          "auth"
        ],
        "description": "REQ-AUTH-004. Issues a 10-minute tempToken.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "pattern": "^\\+[1-9]\\d{6,14}$"
                  },
                  "otp": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  }
                },
                "required": [
                  "phone",
                  "otp"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "tempToken": {
                          "type": "string"
                        },
                        "hasPin": {
                          "type": "boolean"
                        },
                        "isNewUser": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "tempToken",
                        "hasPin",
                        "isNewUser"
                      ],
                      "additionalProperties": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/pin/set": {
      "post": {
        "summary": "Set or reset the 4-digit PIN",
        "tags": [
          "auth"
        ],
        "description": "REQ-AUTH-005. Bearer tempToken required. Creates the user if new; overwrites pinHash if the user already exists (OPEN_QUESTIONS Q2).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pin": {
                    "type": "string",
                    "pattern": "^\\d{4}$"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  }
                },
                "required": [
                  "pin",
                  "name"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string"
                        },
                        "refreshToken": {
                          "type": "string"
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "phone": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string",
                              "enum": [
                                "patient",
                                "provider",
                                "fchv",
                                "admin"
                              ]
                            },
                            "name": {
                              "type": "string"
                            },
                            "facilityId": {
                              "nullable": true,
                              "type": "string"
                            },
                            "facilityName": {
                              "nullable": true,
                              "type": "string"
                            },
                            "createdAt": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "phone",
                            "role",
                            "name",
                            "facilityId",
                            "facilityName",
                            "createdAt"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "accessToken",
                        "refreshToken",
                        "user"
                      ],
                      "additionalProperties": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/pin/login": {
      "post": {
        "summary": "Log in with phone + PIN",
        "tags": [
          "auth"
        ],
        "description": "REQ-AUTH-006. Rate-limited and account-lockout protected.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "pattern": "^\\+[1-9]\\d{6,14}$"
                  },
                  "pin": {
                    "type": "string",
                    "pattern": "^\\d{4}$"
                  }
                },
                "required": [
                  "phone",
                  "pin"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string"
                        },
                        "refreshToken": {
                          "type": "string"
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "phone": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string",
                              "enum": [
                                "patient",
                                "provider",
                                "fchv",
                                "admin"
                              ]
                            },
                            "name": {
                              "type": "string"
                            },
                            "facilityId": {
                              "nullable": true,
                              "type": "string"
                            },
                            "facilityName": {
                              "nullable": true,
                              "type": "string"
                            },
                            "createdAt": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "phone",
                            "role",
                            "name",
                            "facilityId",
                            "facilityName",
                            "createdAt"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "accessToken",
                        "refreshToken",
                        "user"
                      ],
                      "additionalProperties": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Rotate the refresh/access token pair",
        "tags": [
          "auth"
        ],
        "description": "REQ-AUTH-007/011. Reuse of an already-rotated token revokes the whole family.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "refreshToken"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string"
                        },
                        "refreshToken": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "accessToken",
                        "refreshToken"
                      ],
                      "additionalProperties": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/provider/activate": {
      "post": {
        "summary": "Activate provider/fchv role via invite code",
        "tags": [
          "auth"
        ],
        "description": "REQ-AUTH-008. Bearer access token required, role must be patient.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "inviteCode": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "inviteCode"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "phone": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string",
                              "enum": [
                                "patient",
                                "provider",
                                "fchv",
                                "admin"
                              ]
                            },
                            "name": {
                              "type": "string"
                            },
                            "facilityId": {
                              "nullable": true,
                              "type": "string"
                            },
                            "facilityName": {
                              "nullable": true,
                              "type": "string"
                            },
                            "createdAt": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "phone",
                            "role",
                            "name",
                            "facilityId",
                            "facilityName",
                            "createdAt"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "user"
                      ],
                      "additionalProperties": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/me": {
      "get": {
        "summary": "Current authenticated user",
        "tags": [
          "auth"
        ],
        "description": "REQ-AUTH-009. Bearer access token required.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "phone": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string",
                              "enum": [
                                "patient",
                                "provider",
                                "fchv",
                                "admin"
                              ]
                            },
                            "name": {
                              "type": "string"
                            },
                            "facilityId": {
                              "nullable": true,
                              "type": "string"
                            },
                            "facilityName": {
                              "nullable": true,
                              "type": "string"
                            },
                            "createdAt": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "phone",
                            "role",
                            "name",
                            "facilityId",
                            "facilityName",
                            "createdAt"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "user"
                      ],
                      "additionalProperties": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}